break down useSwapTransaction into individual hooks
