+++ src/exchange/swap/hooks/index.ts
@@ -4,3 +4,7 @@ export * from "./usePollKYCStatus";
 export * from "./useSelectableCurrencies";
 export * from "./useSwapProviders";
 export * from "./useSwapTransaction";
+export * from "./useUpdateMaxAmount";
+export * from "./useProviderRates";
+export * from "./useFromState";
+export * from "./useToState";
