+++ src/exchange/swap/hooks/useSwapTransaction.ts
@@ -1,31 +1,27 @@
-import { useState, useReducer, useMemo, useEffect, useCallback } from "react";
+import { useState, useMemo, useCallback } from "react";
 import { BigNumber } from "bignumber.js";
 import useBridgeTransaction, {
   Result as UseBridgeTransactionReturnType,
 } from "../../../bridge/useBridgeTransaction";
-import getExchangeRates from "../getExchangeRates";
-import { getAccountBridge } from "../../../bridge";
-import { getAbandonSeedAddress } from "../../../currencies";
-import { getAccountCurrency, getMainAccount } from "../../../account";
 import type {
   Account,
   TokenAccount,
   TokenCurrency,
   CryptoCurrency,
-  AccountLike,
 } from "../../../types";
 import { ExchangeRate } from "../types";
 import { AmountRequired } from "@ledgerhq/errors";
-import { flattenAccounts } from "../../../account/helpers";
-import { pickExchangeRate, getAccountTuplesForCurrency } from "../utils";
+import { useUpdateMaxAmount } from "./useUpdateMaxAmount";
+import {
+  RatesReducerState,
+  useFromState,
+  useProviderRates,
+  useToState,
+} from ".";
+import { useReverseAccounts } from "./useReverseAccounts";
 
 const ZERO = new BigNumber(0);
 
-export type RatesReducerState = {
-  status?: string | null;
-  value?: ExchangeRate[];
-  error?: Error;
-};
 export type SwapSelectorStateType = {
   currency: null | undefined | TokenCurrency | CryptoCurrency;
   account: null | undefined | Account | TokenAccount;
@@ -41,27 +37,11 @@ export type SwapDataType = {
   refetchRates: () => void;
   targetAccounts?: Account[];
 };
-
-const selectorStateDefaultValues = {
+export const selectorStateDefaultValues = {
   currency: null,
   account: null,
   parentAccount: null,
   amount: null,
-};
-
-const ratesReducerInitialState: RatesReducerState = {};
-const ratesReducer = (state: RatesReducerState, action): RatesReducerState => {
-  switch (action.type) {
-    case "set":
-      return { value: action.payload };
-    case "idle":
-      return { ...state, status: null };
-    case "loading":
-      return { ...state, status: "loading" };
-    case "error":
-      return { status: "error", error: action.payload };
-  }
-  return state;
 };
 
 export type SwapTransactionType = UseBridgeTransactionReturnType & {
@@ -80,62 +60,51 @@ export type SwapTransactionType = UseBridgeTransactionReturnType & {
   fromAmountError?: Error;
 };
 
+export type OnNoRatesCallback = (arg: {
+  fromState: SwapSelectorStateType;
+  toState: SwapSelectorStateType;
+}) => void;
+export type SetExchangeRateCallback = (
+  exchangeRate?: ExchangeRate | null
+) => void;
+
 export const useSwapTransaction = ({
   accounts,
   exchangeRate,
   setExchangeRate,
-  defaultCurrency,
-  defaultAccount,
-  defaultParentAccount,
+  defaultCurrency = selectorStateDefaultValues.currency,
+  defaultAccount = selectorStateDefaultValues.account,
+  defaultParentAccount = selectorStateDefaultValues.parentAccount,
   onNoRates,
 }: {
   accounts?: Account[];
   exchangeRate?: ExchangeRate;
-  setExchangeRate?: (exchangeRate?: ExchangeRate | null) => void;
+  setExchangeRate?: SetExchangeRateCallback;
   defaultCurrency?: SwapSelectorStateType["currency"];
   defaultAccount?: SwapSelectorStateType["account"];
   defaultParentAccount?: SwapSelectorStateType["parentAccount"];
-  onNoRates?: ({
-    fromState,
-    toState,
-  }: {
-    fromState: SwapSelectorStateType;
-    toState: SwapSelectorStateType;
-  }) => void;
+  onNoRates?: OnNoRatesCallback;
 } = {}): SwapTransactionType => {
-  const [toState, setToState] = useState<SwapSelectorStateType>(
-    selectorStateDefaultValues
-  );
-  const [fromState, setFromState] = useState<SwapSelectorStateType>({
-    ...selectorStateDefaultValues,
-    currency: defaultCurrency ?? selectorStateDefaultValues.currency,
-    account: defaultAccount ?? selectorStateDefaultValues.account,
-    parentAccount:
-      defaultParentAccount ?? selectorStateDefaultValues.parentAccount,
-  });
   const [isMaxEnabled, setMax] = useState<SwapDataType["isMaxEnabled"]>(false);
-  const [rates, dispatchRates] = useReducer(
-    ratesReducer,
-    ratesReducerInitialState
-  );
-  const [getRatesDependency, setGetRatesDependency] = useState<unknown | null>(
-    null
-  );
-  const refetchRates = useCallback(() => setGetRatesDependency({}), []);
   const bridgeTransaction = useBridgeTransaction(() => ({
-    account: fromState.account,
-    parentAccount: fromState.parentAccount,
+    account: defaultAccount,
+    parentAccount: defaultParentAccount,
   }));
+  const { fromState, setFromAccount, setFromAmount } = useFromState({
+    accounts,
+    defaultCurrency,
+    defaultAccount,
+    defaultParentAccount,
+    bridgeTransaction,
+  });
+  const { toState, setToAccount, setToAmount, setToCurrency, targetAccounts } =
+    useToState({ accounts });
   const {
     account: fromAccount,
     parentAccount: fromParentAccount,
-    amount: fromAmount,
+    currency: fromCurrency,
   } = fromState;
-  const {
-    account: toAccount,
-    parentAccount: toParentAccount,
-    currency: toCurrency,
-  } = toState;
+  const { account: toAccount } = toState;
   const transaction = bridgeTransaction?.transaction;
   const fromAmountError = useMemo(() => {
     const [error] = [
@@ -150,198 +119,16 @@ export const useSwapTransaction = ({
     bridgeTransaction.status.errors?.gasPrice,
     bridgeTransaction.status.errors?.amount,
   ]);
-  const isSwapReversable = useMemo(() => {
-    if (!toState.account || !fromState.currency) return false;
 
-    const allAccounstWithSub = accounts ? flattenAccounts(accounts) : [];
-    const isToSwappable = !!allAccounstWithSub.find(
-      (account) => account.id === toState.account?.id
-    );
-
-    return isToSwappable;
-  }, [toState.account, fromState.currency, accounts]);
-
-  /* UPDATE from account */
-  const setFromAccount: SwapTransactionType["setFromAccount"] = useCallback(
-    (account) => {
-      const parentAccount =
-        account?.type !== "Account"
-          ? accounts?.find((a) => a.id === account?.parentId)
-          : null;
-      const currency = getAccountCurrency(account as AccountLike);
-
-      bridgeTransaction.setAccount(account as AccountLike, parentAccount);
-      setFromState({
-        ...selectorStateDefaultValues,
-        currency,
-        account,
-        parentAccount,
-      });
-
-      /* @DEV: That populates fake seed. This is required to use Transaction object */
-      const mainAccount = getMainAccount(account as AccountLike, parentAccount);
-      const mainCurrency = getAccountCurrency(mainAccount);
-      const recipient = getAbandonSeedAddress(mainCurrency.id);
-      bridgeTransaction.updateTransaction((transaction) => ({
-        ...transaction,
-        recipient,
-      }));
-    },
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [accounts, bridgeTransaction.updateTransaction]
-  );
-
-  /* UPDATE to accounts */
-  const setToAccount: SwapTransactionType["setToAccount"] = useCallback(
-    (currency, account, parentAccount) =>
-      setToState({
-        ...selectorStateDefaultValues,
-        currency,
-        account,
-        parentAccount,
-      }),
-    []
-  );
-
-  /* Get the list of possible target accounts given the target currency. */
-  const getTargetAccountsPairs = useCallback(
-    (currency) =>
-      currency &&
-      accounts &&
-      getAccountTuplesForCurrency(currency, accounts, false),
-    [accounts]
-  );
-  const targetAccounts = useMemo(
-    () =>
-      getTargetAccountsPairs(toCurrency)?.map(
-        ({ account, subAccount }) => subAccount || account
-      ),
-    [toCurrency, getTargetAccountsPairs]
-  );
-
-  const setToCurrency: SwapTransactionType["setToCurrency"] = useCallback(
-    (currency) => {
-      const targetAccountsPairs = getTargetAccountsPairs(currency);
-      const accountPair = targetAccountsPairs && targetAccountsPairs[0];
-      const account =
-        accountPair && (accountPair.subAccount || accountPair.account);
-      const parentAccount =
-        accountPair && accountPair.subAccount && accountPair.account;
-
-      setToState({
-        ...selectorStateDefaultValues,
-        currency,
-        account,
-        parentAccount,
-      });
-    },
-    [getTargetAccountsPairs]
-  );
-
-  const setFromAmount: SwapTransactionType["setFromAmount"] = useCallback(
-    (amount) => {
-      bridgeTransaction.updateTransaction((transaction) => ({
-        ...transaction,
-        amount,
-      }));
-      setFromState((previousState) => ({ ...previousState, amount: amount }));
-    },
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [bridgeTransaction.updateTransaction]
-  );
-
-  const setToAmount: SwapTransactionType["setToAmount"] = useCallback(
-    (amount) =>
-      setToState((previousState) => ({ ...previousState, amount: amount })),
-    []
-  );
-
-  /* UPDATE from amount to the estimate max spendable on account
-  change when the amount feature is enabled */
-  useEffect(() => {
-    const updateAmountUsingMax = async () => {
-      if (!bridgeTransaction.account) return;
-      const bridge = getAccountBridge(
-        bridgeTransaction.account,
-        bridgeTransaction.parentAccount
-      );
-      const amount = await bridge.estimateMaxSpendable({
-        account: bridgeTransaction.account,
-        parentAccount: bridgeTransaction.parentAccount,
-        transaction: bridgeTransaction?.transaction,
-      });
-      setFromAmount(amount);
-    };
-
-    if (isMaxEnabled) {
-      updateAmountUsingMax();
-    }
-  }, [
-    setFromAmount,
-    isMaxEnabled,
-    fromState.account,
-    bridgeTransaction?.transaction?.feesStrategy,
-  ]);
-
-  /* Fetch and update provider rates. */
-  useEffect(() => {
-    let abort = false;
-    async function getRates() {
-      if (
-        !transaction ||
-        !transaction?.amount ||
-        !transaction?.amount.gt(0) ||
-        !toAccount ||
-        !fromAccount ||
-        getAccountCurrency(toAccount) !== toCurrency
-      ) {
-        setExchangeRate && setExchangeRate(null);
-        return dispatchRates({ type: "set", payload: [] });
-      }
-      dispatchRates({ type: "loading" });
-      try {
-        let rates: ExchangeRate[] = await getExchangeRates(
-          { fromAccount, fromParentAccount, toAccount, toParentAccount },
-          transaction
-        );
-        if (abort) return;
-        if (rates.length === 0) {
-          onNoRates && onNoRates({ fromState, toState });
-        }
-        // Discard bad provider rates
-        let rateError: Error | null | undefined = null;
-        rates = rates.reduce<ExchangeRate[]>((acc, rate) => {
-          rateError = rateError ?? rate.error;
-          return rate.error ? acc : [...acc, rate];
-        }, []);
-        if (rates.length === 0 && rateError) {
-          // If all the rates are in error
-          dispatchRates({ type: "error", payload: rateError });
-        } else {
-          dispatchRates({ type: "set", payload: rates });
-          setExchangeRate &&
-            exchangeRate &&
-            pickExchangeRate(rates, exchangeRate, setExchangeRate);
-        }
-      } catch (error) {
-        !abort && dispatchRates({ type: "error", payload: error });
-      }
-    }
-
-    getRates();
-
-    return () => {
-      abort = true;
-      dispatchRates({ type: "idle" });
-    };
-  }, [
+  const { isSwapReversable, reverseSwap } = useReverseAccounts({
+    accounts,
     fromAccount,
-    fromAmount,
     toAccount,
-    transaction,
-    getRatesDependency,
-    onNoRates,
-  ]);
+    fromParentAccount,
+    fromCurrency,
+    setFromAccount,
+    setToAccount,
+  });
 
   const toggleMax: SwapTransactionType["toggleMax"] = useCallback(
     () =>
@@ -354,27 +141,35 @@ export const useSwapTransaction = ({
     [setFromAmount]
   );
 
-  const reverseSwap: SwapTransactionType["reverseSwap"] = useCallback(() => {
-    if (isSwapReversable === false) return;
-
-    const [newTo, newFrom] = [fromState, toState];
-    setFromAccount(newFrom.account);
-    setToAccount(newTo.currency, newTo.account, newTo.parentAccount);
-  }, [fromState, toState, setFromAccount, setToAccount, isSwapReversable]);
-
-  const swap = {
-    to: toState,
-    from: fromState,
+  useUpdateMaxAmount({
+    setFromAmount,
     isMaxEnabled,
-    isSwapReversable,
-    rates,
-    refetchRates,
-    targetAccounts,
-  };
+    account: fromAccount,
+    parentAccount: fromParentAccount,
+    transaction,
+    feesStrategy: transaction?.feesStrategy,
+  });
+
+  const { rates, refetchRates } = useProviderRates({
+    fromState,
+    toState,
+    exchangeRate,
+    transaction,
+    onNoRates,
+    setExchangeRate,
+  });
 
   return {
     ...bridgeTransaction,
-    swap,
+    swap: {
+      to: toState,
+      from: fromState,
+      isMaxEnabled,
+      isSwapReversable,
+      rates,
+      refetchRates,
+      targetAccounts,
+    },
     setFromAmount,
     toggleMax,
     fromAmountError,
